requirements.txt
Flask==3.0.3
google-cloud-aiplatform>=1.49.0
python-dotenv>=1.0.0
gunicorn==21.2.0
